rename read_rows -> readrows, fix docs
